+++ src/Practitioner/PatientProfile/Header.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import useStores from '../../Basics/UseStores';
 import { observer } from 'mobx-react';
-import Buttons from './MobileView/Buttons'
+import Buttons from './Buttons'
 import Avatar from '@material-ui/core/Avatar';
 import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography';
